+++ packages/leancode_contracts_generator/lib/src/contracts_generator_config.dart
@@ -92,8 +92,6 @@ input:
   #     - glob/not_you.cs
   #   # The base directory used for globbing
   #   directory: .
-  # # Overrides the default backend generator version. Can lead to unexpected outputs.
-  # override_generator_version: "${GeneratorScript.supportedVersion}"
 # Name of the output dart file
 name: contracts
 # Regex to filter-in namespaced statements
@@ -106,19 +104,12 @@ directives: ""
 extra: ""
 ''';
 
-  static String? _getOverridenGeneratorVersion(dynamic config) {
-    return config['override_generator_version'] as String?;
-  }
-
   static GeneratorScript? _configureProject(dynamic config) {
     if (config['project'] == null) {
       return null;
     }
 
-    return GeneratorScript.project(
-      (config['project'] as List).cast<String>(),
-      overrideGeneratorVersion: _getOverridenGeneratorVersion(config),
-    );
+    return GeneratorScript.project((config['project'] as List).cast<String>());
   }
 
   static GeneratorScript? _configurePath(dynamic config) {
@@ -131,7 +122,6 @@ extra: ""
       (values['include'] as List).cast<String>(),
       exclude: (values['exclude'] as List?)?.cast<String>(),
       directory: values['directory'] as String?,
-      overrideGeneratorVersion: _getOverridenGeneratorVersion(config),
     );
   }