+++ packages/leancode_contracts_generator/lib/src/generator_script.dart
@@ -2,22 +2,17 @@ import 'dart:io';
 
 import 'contracts_generator_exception.dart';
 
-part 'generator_script.g.dart';
-
 /// Interface to the C# protobuf generator
 class GeneratorScript {
   /// Generate for a list of projects. Passed paths need to point to a .csproj file.
-  GeneratorScript.project(
-    List<String> projects, {
-    this.overrideGeneratorVersion,
-  }) : args = ['project', '-p', ...projects];
+  GeneratorScript.project(List<String> projects)
+      : args = ['project', '-p', ...projects];
 
   /// Generate for a all files in the globbed paths
   GeneratorScript.path(
     List<String> include, {
     List<String>? exclude,
     String? directory,
-    this.overrideGeneratorVersion,
   }) : args = [
           'path',
           '-i',
@@ -33,21 +28,14 @@ class GeneratorScript {
         ];
 
   final List<String> args;
-  final String? overrideGeneratorVersion;
-
-  static const supportedVersion = '3.0.0';
-
-  static const String _script = _$script;
 
   Future<List<int>> run() async {
+    await Process.run('dotnet', ['tool', 'restore']);
     final result = await Process.run(
-      'sh',
-      ['-c', _script, '--', ...args, '-o-'],
+      'dotnet',
+      ['tool', 'run', 'dotnet-contracts-generate', '--', ...args, '-o-'],
       stdoutEncoding: null,
       stderrEncoding: null,
-      environment: {
-        'SERVER_VERSION': overrideGeneratorVersion ?? supportedVersion,
-      },
     );
 
     if (result.exitCode != 0) {