+++ packages/leancode_contracts_generator/scripts/generate_files.sh
@@ -4,11 +4,3 @@
 
 protoc -I=contractsgenerator --dart_out=lib/src/proto contractsgenerator/contracts.proto
 dart format lib/src/proto/*.dart
-
-
-echo "\
-part of 'generator_script.dart';
-
-const _\$script = r'''
-`cat contractsgenerator/tools/generate.sh`
-''';" > lib/src/generator_script.g.dart